+++ preset/interceptors/commonThrottleRequest.js
@@ -29,75 +29,75 @@
 export default function commonThrottleRequest(throttleFilter, option) {
   const defaultOption = {
     notInterceptKey: 'notThrottle',
-  }
-  option = Object.assign({}, option, defaultOption)
+  };
+  option = Object.assign({}, defaultOption, option);
   if (!throttleFilter && typeof throttleFilter !== 'function') {
-    throttleFilter = (store, now) => store.url === now.url && store.config.method === now.config.method
+    throttleFilter = (store, now) => store.url === now.url && store.config.method === now.config.method;
   }
 
   return {
     install(interceptors) {
       /** @type {RequestInfo[]} 保存*/
-      let cacheArr = []
+      let cacheArr = [];
       let cacheArr_clean = config => {
         // 请求已返回则移除保存
         for (let i = 0; i < cacheArr.length; i++) {
           if (cacheArr[i]?.requestId === config._commonThrottleRequest?.requestId) {
-            cacheArr[i] = null
-            break
+            cacheArr[i] = null;
+            break;
           }
         }
         if (cacheArr.length > 20) {
-          cacheArr = cacheArr.filter(Boolean)
+          cacheArr = cacheArr.filter(Boolean);
         }
-      }
-      let requestId = 1
+      };
+      let requestId = 1;
 
       interceptors.request.use(
         (url, config) => {
-          if (config[option.notInterceptKey]) return config
+          if (config[option.notInterceptKey]) return config;
 
-          let hasRequestStored = false
-          let emptyIndex = cacheArr.length // cacheArr 中空位的index
-          const storeObj = { requestId, url, config }
+          let hasRequestStored = false;
+          let emptyIndex = cacheArr.length; // cacheArr 中空位的index
+          const storeObj = { requestId, url, config };
           for (let i = 0; i < cacheArr.length; i++) {
-            const item = cacheArr[i]
+            const item = cacheArr[i];
             if (!item) {
-              emptyIndex = i
-              continue
+              emptyIndex = i;
+              continue;
             }
             if (throttleFilter(item, { url, config })) {
-              hasRequestStored = true
-              throw new Error('commonThrottleRequest: 该请求已发起未返回，不能重新发起。已忽略。')
+              hasRequestStored = true;
+              throw new Error('commonThrottleRequest: 该请求已发起未返回，不能重新发起。已忽略。');
             }
           }
           // 传递到response 回调中
           config._commonThrottleRequest = {
             requestId: requestId++,
-          }
+          };
 
           if (!hasRequestStored) {
             // 没保存请求则保存
-            cacheArr[emptyIndex] = storeObj
+            cacheArr[emptyIndex] = storeObj;
           }
-          return config
+          return config;
         },
         err => {
-          if (err.config) cacheArr_clean(err.config)
-          return Promise.reject(err)
+          if (err.config) cacheArr_clean(err.config);
+          return Promise.reject(err);
         },
-      )
+      );
 
       interceptors.response.use(
         (data, { config }) => {
-          cacheArr_clean(config)
-          return data
+          cacheArr_clean(config);
+          return data;
         },
         err => {
-          if (err.config) cacheArr_clean(err.config)
-          return Promise.reject(err)
+          if (err.config) cacheArr_clean(err.config);
+          return Promise.reject(err);
         },
-      )
+      );
     },
-  }
+  };
 }